+++ LunarDisturbances/Integrations/IContentPatcherAPI.cs
@@ -1,0 +1,10 @@
+﻿using StardewModdingAPI;
+
+namespace TwilightShards.LunarDisturbances.Integrations
+{
+    /// <summary>The Content Patcher API which other mods can access.</summary>
+    public interface IContentPatcherAPI
+    {
+        void RegisterToken(IManifest mod, string name, object token);
+    }
+}